requirements.txt
python-telegram-bot[rate-limiter]==20.1
openai>=0.27.0
tiktoken>=0.3.0
PyYAML==6.0
pymongo==4.3.3
python-dotenv==0.21.0
pydub==0.25.1


langchain~=0.0.171
yaml~=0.2.5
cohere~=4.5.1
pyrogram~=2.0.106